Explore the architecture; make it mine
Most notably, I modified the Todo views to no longer call the api
directly. Instead, they call an action creator. The action creator is
responsible for making API calls, as shown in the Flux picture that's
included on the home page.

I also prefer JSX syntax, so I changed things to that.